requirements.txt
Flask==2.3.3
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
matplotlib==3.7.2
weasyprint==60.0
python-dateutil==2.8.2
